Allow queries on collections setting conditions for more then one item
This example of modification on `Query::equals` would allow making queries on multiple items in a collection.
Let's say you have an entity Game with a collection of GamePlayer(s) and you search for a Game with a player1 in position forward and player2 in position defender, the actual flow Query would not allow it, with the code change you can construct the constraint this way:
```
$query->logicalAnd(
    $query->equals('gamePlayer.player', $player1, true, 'p1'),
    $query->equals('gamePlayer.position', 'forward', true, 'p1'),
    $query->equals('gamePlayer.player', $player2, true, 'p2'),
    $query->equals('gamePlayer.position', 'defender', true, 'p2')
);

Although the change is trivial, there maybe consequences for the QueryInterface and others that I'm anware of. Full implementation of the change would require updating all methods getPropertyNameWithAlias is used and proper tests.